Add emoji support through dependency
This can be enabled / disabled by changing the
`python.slack_extension.replace_emoji` variable in `plugins.conf`.

I have a solution that would not require a dependency:

```python
EMOJI_DICT = [
    (':+1:', u'\U0001F44D'),
    (':grinning:', u'\U0001f600'),
    (':loudspeaker:', u'\U0001f4e2')
]

for k, v in EMOJI_DICT:
    message = message.replace(k, v)
```

However I don't feel it is up to this project to maintain a list of
emoji-conversions, so I figured we could use a third party library.

However we might want to introduce a CUSTOM_EMOJI_DICT that the user can
configure for custom emojis their team have. This list would be
maintained in `plugins.conf`.