Check for bot_id earlier
Some bots have 'username' set for them, which could lead to an
inconsistent experience where some bots would have the robot smiley,
while others did not.  Check for bot_id earlier and use the
formatted_name of the given bot.

Also add a space in front of the robot smiley to make it easier to
see.