Show nicks for reactions
Instead of only showing the number of persons that added a reaction,
show the nicks that added them. The nicks are shown comma separated in
parenthesis after each reaction.

This adds an item named users to the reaction dict of a message, which
contains a list of all of the user ids. Since the code assumes that item
exists, which makes it incompatible with the old structure, the
CACHE_VERSION is increased by one.