Add a config option for showing nicks for reactions
Some may think it takes up too much space to show all the nicks for all
of the reactions. This adds a config option `show_reaction_nicks` for
configuring that. If set to false, it only shows the number of persons
that added each reaction, like before. If set to true, it shows the new
version with each nick. It defaults to false, to not change the
behavior.