Add support for sending a reaction
A reaction can be added to the last message in the channel by writing

    +:reaction:

Where reaction is the name of the reaction. This is the same syntax that
Slacks web interface uses.

Additionally, it supports sending a reaction to an earlier message by
prefixing the text with a number. The number is the number of the
message to add a reaction to, counting from the bottom. E.g. 1 is the
last message (same as without a number), 2 is the one before etc.

This fixes #155.