Fix up displayed user name for new user and new channel messages
Use the value from the config option weechat.look.prefix_network if no
user is passed in. Thanks to @bendem for the suggestion to use this
rather than a hard coded string.

This also fixes a bug where the server buffer was passed as the
message and the user was set to "New user joined: foo".