Prefix the server buffer name with slack instead of suffixing it
Instead of using the format `$team.slack.com` for the server buffer
name, change it to `slack.$team`. This makes it easier to see that the
buffer is a slack buffer.

Additionally, it makes it possible to set options where you specify the
start of the buffer name for all of the slack buffers. For example, it
makes it possible to set the option `logger.mask.python.slack` for
setting a logger file mask for all of the slack buffers.