Return from command_talk if user or channel was not found
Not returning when this error condition is reached causes a traceback
when we try to set the buffer, since no channel will exist for the
specified user (and channel.channel_buffer will be None).

This commit returns ``False`` so that we do not try to set a buffer for
a nonexistant channel.