Make it possible to set timeouts on slack URL comms
On very busy slack networks the default timeout of 20000 can be a
problem. This timeout can mean that the extension is unable to download
all the necesarry data before weechat times out the connection.

This patch allows for the timeout to be adjusted so that for larger
networks the end user can specify it.

Based on #279 by @GregBowyer.

Closes: #279