Use unicode literals everywhere
This changes all strings to be unicode strings in an attempt to handle
unicode better. The idea is to use unicode everywhere internally in
wee_slack and decode/encode where data is received/sent to/from
wee-slack.

That means we have to decode strings received from weechat or slack and
encode strings we send to them. This way encoding and decoding is done
on specific places, and if done everywhere, we should not get any
unicode errors.