Create a wrapper around weechat which handles encoding
This encodes everything sent into weechat, and decodes the return
values, so we don't have to do it every place we call weechat. We still
have to decode the arguments in all of the callback functions though.

The encode_to_utf8 and decode_from_utf8 functions traverses data
structures and encodes/decodes all of the strings inside the structures.