Fix channel history not appearing when an earlier message has 'text': None
Today I learned that sometimes, rather than the key missing, it has the value
None. In this case, (key in dict) is true!

This fixes an issue where, if the backscroll of a channel contained such a
message, it would throw while fetching the backscroll and cut off without
fetching all of it.

Signed-off-by: Ben Kelly <btk@google.com>
Signed-off-by: Ben Kelly <bk@ancilla.ca>