Fix #118: Add command to compose multiline message
Borrows heavily from #365.

This adds a `/slack edit [extension] [channel]` command that allows
editing multiline messages in the external editor of choice. If
[extension] is passed, it will be used as the extension of the temporary
file to edit the message, in this mode code backticks are added
implicitly and hence are not needed to be inserted manually. Otherwise,
the default `md` file extension (for markdown) is assumed and no
backticks are automatically generated.

The [channel] parameter can be used to direct a message to a given
channel.