Prevent lines being broken in bare display mode
If you update a line with hdata_update and pass a string containing
newlines, the buffer will become a bit broken when bare display mode is
activated (alt+l). It seems that the newlines are printed in bare mode,
but extra lines are not inserted, so the lines are overlapping with each
other.

Fixes #306