test: Make mock_weechat a dependency of realish_eventrouter
Setting up the eventrouter for the tests involves code that requires
the config object to be set, so it has to depend on mock_weechat.

This didn't cause a problem when all the tests were run since the first
test depends on mock_weechat so it would be set up. However, if you
tried to run a single test that depended on realish_eventrouter, but not
mock_weechat, it would fail.