test: Empty eventrouter queue when initializing it
The commit 6718e2f added some more events to the queue when a buffer is
created. This made some tests fail as it would process these events
instead of the ones added in the tests. Fix it by processing all events
in the queue when setting up the eventrouter for the tests.