Add completion of custom emojis
This loads the list of custom emojis for each team after wee-slack has
connected to the team, and adds them to the emoji completion list for
that team.

The list of custom emojis is only loaded if the list of the standard
emojis has been loaded, since it would be a bit strange to only have
completions of the custom emojis.