Ensure that EMOJI is always defined
After the last commit, the EMOJI variable is accessed in the
initialization of SlackTeam. This made some tests fail since they don't
run the main code, so it would not be defined. By setting EMOJI to an
empty list in the top, we ensure that this is not a problem.