fix: Set new_messages in set_unread_count_display
The new_messages property is used by mark_read to determine if it should
excecute. When a channel is initially loaded, new_messages is set to
False, and unread_count_display is set to the number of unread messages.
new_messages is not set to True until a new message arrives.

This means that if you call mark_read after loading a channel, but
before a new message has arrived, it won't do anything, even if the
channel had unread messages when it was loaded. By setting new_messages
in set_unread_count_display, this is fixed.