Hide activity from muted channels
This sets the correct tags for messages in muted channels and prevents
muted channel from being added to the hotlist on start and when a
message arrives.

Threads from a muted channel are not muted. If you are part of a thread,
you probably want to be notified for messages in it, even though the
parent channel is muted. The official client shows messages in threads
from muted channels in "New threads" with a highlight in the sidebar.

Fixes #456