Reload emoji list on emoji change event
It's simpler to just reload the whole list instead of adding and
removing specific emojis, especially since we would have to remove
aliases for an emoji when the emoji is removed (and we don't store if an
emoji is an alias or not).

Since this probably doesn't happen that often, it shouldn't be a problem
to do it this way.