Fix typing indicator when having multiple windows
If you have multiple windows in weechat, you get (by default) one typing
bar per window. Previously, the bar used the globally active channel
(the active channel in the active window) for all the bars, so all the
bars would show the same info. This changes it to use the active channel
for the window the bar is in.

The callback name is changed to include (extra), as that is necessary
for getting the active buffer for each window as an argument to the
callback function.