Replace Slack links with direct links to shared files
The current links in shared file messages only show a thumbnail and
require an extra click to get to the actual shared file, which is a
minor annoyance.

Links from external members in shared channels are also from their
team's Slack subdomain, so they're inaccessible if you're not a member
of their team. The direct links are from the files.slack.com subdomain
though and visible to all, and the official Slack client uses them to
show those files.