fix: corrected registration instructions
in the registration dialog, wee_slack asks the user to reload the
slack script via /script. in my weechat 2.1 on debian sid, this is
not the proper way to reload scripts anymore. instead one has to
use the /python command to achieve the same effect. this commit
corrects the misleading instruction message.