Combine users and external_users list
Since we have an is_external attribute on the users, it's simpler to
just have them all in a single list.

The documentation says that the user id is not necessarily unique across
teams, but I talked to Slack and they said it's unlikely to be a
collision. If you refer to a user in a message, you only use the id
without a team id, so you will have problems with collisions there
anyway. I'm waiting to hear from Slack how to handle that.