Render files after message instead of changing message text
Instead of changing the message text, append the files belonging to the
message when we render the message. This is the same as we do with
attachments. This fixes the issue where the files would not be shown
anymore if a message is changed.

The rendering format of the files is changed to match the way we render
refs.