Move message edited text to render and remove message suffix
The message suffix is only used to show that a message is edited.
Instead of keeping track of the suffix, just check if the message has
been edited in the render function, and add the text there. This way, we
only have to add the '(edited)' text one place.

With this change, the edited text moves to right after the message text,
before the attachments and files. It also fixes a bug where '(edited)'
would not show on messages that have a thread.