Clear messages dict when refetching history
When refetching the history, we clear the buffer because the messages
will be printed again. Now the dicts for messages and hashed_messages
are also cleared, so they match what we actually have in the buffer.

This is necessary so we can check if a message is in the messages dict
to find out if it's printed to the buffer, which we need for the next
commit.