Don't process message again if it already exists
The API has started to send both responses for sent messages as well as
message events for the same messages. Previously, we would only get
message events for messages from other clients. For the messages we send
ourselves we would just get responses, and no message events.

This caused wee-slack to print own messages twice, once for the response
and once for the message event. To prevent this, check if we already
have a message for the ts, and if we do, ignore the new message.

I don't want to rely on only the message events and ignore the
responses, because this change has rolled out gradually, so not all
users might get message events for own messages yet. Additionally, the
docs are not clear on what the behavior should be.

Fixes #606