Clear messages before processing history
If a message arrived after history was requested, but before the
response arrived, it would be lost. It would be printed after the
"getting channel history" text, and then handle_history would clear
that, and it would not be printed again since the message ts already was
processed.

By clearing the messages dict, the message will get processed and
printed again. So we should never clear the buffer without also clearing
the messages, to prevent such problems.