Don't render messages surrounded by _ as actions
Even though me_messages and messages surrounded by underlines render
equally in the official client, it is two different types of messages.
Converting messages surrounded by underlines to me_messages causes
problems when editing, because the underlines will be removed from the
message so it won't be shown as an action anymore.

The official clients sends me_messages as that subtype, and after the
previous commit, wee-slack also does this, so there is no reason for
this conversion anymore.

Since subtype isn't changed in SlackMessage anymore, we can use the
subtype variable we have, and the comment is not needed anymore.