Set tag slack_muted_channel on muted channel messages
The tag slack_muted_channel is set on all messages in channels which are
muted, and all messages in threads which belong to muted channels.

If you use a plugin for push notifications and don't want notifications
from muted channels, you can set it to ignore messages with this tag.
This would give you the same behavior as with the official client. Note
that this just applies to messages shown in the hotlist, i.e. highlights
(depending on your config) and thread messages, as messages hidden from
the hotlist are probably already ignored by the notification plugin.