Group color configs together
Renames the config options for colors to start with color instead of end
with it, so they are grouped together.

If thread_suffix_color is set, the config will be migrated to the new
name. I didn't bother to migrate reaction_suffix_color since that added
just half an hour ago.