Fix bug preventing first line in buffer to be changed
This rewrites modify_buffer_line and modify_print_time (now
modify_last_print_time to be more clear) to be much more readable.

modify_buffer_line had a bug where it moved past the first line, causing
the line_pointer to be lost so the first line couldn't be updated. This
is now fixed.