Use buffer from arguments instead of w.current_buffer
The buffer given in the arguments refers to the buffer where the command
was run, while w.current_buffer refers to the buffer active in weechat.
Normally these are the same, but relay clients may have a different
buffer active than the one in weechat. Therefore, using w.current_buffer
could cause commands sent from those clients to operate on the wrong
buffer.

Fixes #547 and #615