Add custom commands properly
We should use hook_command instead of hook_command_run when we add new
custom commands because this will include completion and help for the
command.

When we use this, the command name won't be included in the arguments,
which is why the usage of args in the functions has changed.