Add show_buflist_presence config option (#558)
This commit adds a `show_presence_presence` config option that defaults to
true.  When true, the `+` appears before present users in the buffer
list, as they did before this commit.  But when false, the `+` mark is
omitted and presence is no longer shown.

This helps users who want their terminal programs to flag activity but
don't consider presence changes to be activity worth flagging.