Move message handling out of exception handling
Remove handling of all exceptions. If an unknown exception happens, it's
better to just let it print a stack trace to the core buffer so we
notice it, instead of just printing it in the debug buffer.

The two specific exceptions we handle are just necessary for the recv
call, so we can move the rest of the lines outside of the exception
handling.