Process the entire websocket queue in receive_ws_callback
When we receive many lines over the websocket, receive_ws_callback seems
to be called periodically (10 times per second or so) until all lines
have been read. Since we previously only read one line per invocation of
receive_ws_callback, it would take a long time to read all the lines.

In addition to making events processed slower than necessary, this
caused a problem for the pong watcher. Since the lines was processed so
slow, it might take a long time until we process the pong message when
the queue was large. This caused the script to disconnect the websocket
since it didn't see the pong messages in time.

To fix this, we run websocket receive in receive_ws_callback until there
are no more messages. The messages are just added to the eventrouter
queue, not processed right away, so this shouldn't cause performance
issues.