Only check for pong responses when ping has been recently sent
With large teams on a slow computer, wee-slack may unfortunately hang
for a while at times while processing data. When this happen, it doesn't
get a chance to send pings. When it's finished processing, it previously
might have disconnected if the pong checker runs before sending a ping,
since it's a long time since a ping has been sent.