Fix broken receive_httprequest_callback when record_events is true
When removing pickle in commit 3639342 I broke record_events without
realizing it. record_events tries to json serialize the response dict,
and that failed because that dict includes some metadata which without
pickle is a class instead of a string.

Fix it by recording the event before adding the metadata. I don't think
there is much point in recording the metadata, and the serialized output
from pickle wasn't practical anyways.