Use hook_process_hashtable for uploading files
Instead of creating a space separated string of the arguments for curl,
create a dict and use hook_process_hashtable. The benefit of this is
that we don't need to escape the spaces in the file name.

This also utilizes SlackRequest to create the url we post to, so we
don't have to hard code the url in this function.

I tried to use hook_process_hashtable to request the url directly,
instead of calling the curl command, but I didn't manage to make it
upload the file in a format accepted by Slack. I opened an issue to
weechat about it: https://github.com/weechat/weechat/issues/1352