Remove escape characters from path if file is not found
Some terminals (at least Terminal.app and iTerm2) escape spaces in the
path when you drop a file in it. However, weechats tab completion does
not escape them, so we can't just remove the escape characters, in case
a user is trying to upload a file which has a name that contains "\ ".

Therefore, first check if the file exists, and then remove escapes if it
doesn't, and check if the unescaped name exists.

Fixes #360