Prevent websocket-client from importing and using numpy
This prevents a crash when you reload the script when using python 3 if
numpy is installed. When weechat unloads a script it calls Py_Finalize,
and when it loads a script, it calls Py_Initialize. If numpy is
imported, this causes python to segfault, which crashes weechat.

By setting numpy in sys.modules to None, it won't be imported when
websocket-client tries to import it. numpy is an optional dependency of
websocket-client, and if it's not installed, it will set numpy to None
itself, so therefore, this should work just like when numpy isn't
installed.

See https://github.com/numpy/numpy/issues/11925 for more details about
the issue in numpy.