new_messages should not be set for backlog messages, but for self_msgs
There are scenarios where last_read isn't set, and only self messages
are printed. I experienced it when the only message in the channel was
me joining and renaming the channel. In that case, last_read were never
set, so we should set new_messages for our own messages so last_read
will be set.