Linkify users in /slack slash command
If we don't linkify users, only usernames (which are not shown anymore)
are accepted by the chat.command endpoint, not display names. It does
accept user links, so we can use that. However, it doesn't seem to
accept channel links (at least the invite command, which is what I
tested), so we only linkify users.