Fallback to full name if display name is not set
The username can't be changed by the users themselves now and is auto
generated for new users, so we shouldn't use it. Only use username as a
last resort if neither display name nor full name is set (I'm not sure
this ever happens, it looks like the full name field is mandatory).