Add basic support for private channels converted from public
This is for channels which initially were public, but then are changed
to private. They look the same as groups (channels initially created as
private) in the official clients, but can only be used with the
conversations api. They are listed in the channels list of rtm.start
with is_private set to true.

There are some remaining issues, which are documented in the readme.
I'm not sure if we can support read sync and showing unread on load
without changes from Slack to the API.

Showing thread messages in the channel is possible to fix, but we would
have to send requests to load the thread history for all the messages in
the history that has replies.

Fixes most of #644