Fetch members explicitly for SlackPrivateChannel
Like SlackSharedChannel, SlackPrivateChannel doesn't include members in
the result from rtm.start, and we have to use conversations.info for
them which doesn't include members either. Therefore, we have to call
conconversations.members.

I hadn't noticed that the members were missing because I only tested
with channels with a short backlog, so the members were filled by the
join messages which were still present in the history that were fetched.