Set channel.is_member when channel is created
The channel dict sent as kwargs to SlackChannel from handle_rtmstart
contains the is_member property. However the one from
process_channel_created doesn't. Since you never are a member when the
channel is created, we can set it to false (if you create the channel
yourself or is invited immediately a channel_joined message will follow
the channel_created message).