Set tags in deterministic order with slack_ts first
Since we look up the message ts from the tags as of the last commit, we
should place the ts tag first so we don't have to loop through the other
tags when finding it. This should make it a bit faster to find a
message, especially when we have to search through many lines.