Show away in away bar item when presence is away
This shows it in the standard away bar item, and also adds a specific
one for wee-slack. The reason for including the specific one is that if
multiple scripts in the same plugin define the same bar item, only one
of them will work[0]. So people that have another python script which
also defines the away bar item can use the slack_away bar item instead.

[0]: https://github.com/weechat/weechat/issues/1348