Strip whitespace from help descriptions for commands
command_reply now has a blank line after the first line, which means we
sent in a description with a blank first. Weechat automatically adds a
blank line before the description, so it showed two blank lines for
/help reply. We want the blank line in the docstring for /slack help
reply, so strip it in the description instead of removing it from the
docstring.