Preserve thread channels across reconnections and /rehistory
Previously, the references to the thread channels and the hashes of the
messages would be lost when the channel reloaded its history. That meant
that when new messages arrived in the thread, they would not be printed
in the thread channel.

This moves the references to the thread channels from each parent
message to the channel, and changes the values in hashed_messages to be
just the timestamp instead of the message, so we don't have to clear it
when we reload the history.

Relates to #714