Don't add fallback if link exists in text_before
We should only use the fallback if we don't get any information out of
the attachment. For links that exists in text_before, we have gotten the
information from the attachment, but choose not to display it so we
don't end up with duplicate links. Therefore, we shouldn't use the
fallback in this case.