Stop calling thread.mark if not supported
The subscriptions.thread.mark method is only supported for the tokens
the web client use, not by oauth or legacy tokens, so if we get an error
that it's not supported for the current token type, remove it from
slack_api_translator so we don't keep calling it. slack_api_translator
is moved to a team attribute because different teams may use different
token types.