Set team in message_json after recording event
SlackTeam can't be json serialized, so this broke record_events. I broke
this in commit 324890c when I changed from storing the team hash on the
message_json to storing the team.

You can't currently see which team the message is from, but you couldn't
really with the team hash either since the relation between the team and
the team hash is not printed anywhere. I will add a better way to see
which team the message is from in the next commit.

Fixes #761