Support toggling a reaction by right clicking on an emoji or emoji name
If you right click on an emoji or an emoji name in a message, that emoji
will be either added as a reaction, or removed if you already have
reacted with it. This allows you to easily react with the same reactions
as others have used. If you click anywhere else on a message, the
message hash is inserted in the input field as before.

For emoji names you have to click on the first colon. This is partly
because it was easier to implement, and partly because I'm going to
trigger this using the keyboard (using https://github.com/trygveaa/kitty-hints-mouse-click),
so I don't have a need for clicking other places on the emoji name.

If you want to keep the old behavior of inserting the message hash in
the input field no matter where you click on the message, change the
binding for right click to hsignal:slack_cursor_message, like this:

    /key bindctxt mouse @chat(python.*):button2 hsignal:slack_cursor_message